+++ A-pex/src/BiCriteria/BOAStar.cpp
@@ -26,7 +26,7 @@ void BOAStar::operator()(size_t source, size_t target, Heuristic &heuristic, Sol
     std::vector<NodePtr> open;
     std::make_heap(open.begin(), open.end(), more_than);
 
-    node = std::make_shared<Node>(source, std::vector<size_t>(2,0), heuristic(source));
+    node = std::make_shared<Node>(source, std::vector<double>(2,0), heuristic(source));
     open.push_back(node);
     std::push_heap(open.begin(), open.end(), more_than);
 
@@ -64,7 +64,7 @@ void BOAStar::operator()(size_t source, size_t target, Heuristic &heuristic, Sol
         const std::vector<Edge> &outgoing_edges = adj_matrix[node->id];
         for(auto p_edge = outgoing_edges.begin(); p_edge != outgoing_edges.end(); p_edge++) {
             size_t next_id = p_edge->target;
-            std::vector<size_t> next_g = {node->g[0]+p_edge->cost[0], node->g[1]+p_edge->cost[1]};
+            std::vector<double> next_g = {node->g[0]+p_edge->cost[0], node->g[1]+p_edge->cost[1]};
             auto next_h = heuristic(next_id);
 
             // Dominance check
