+++ A-pex/src/BiCriteria/ShortestPathHeuristic.cpp
@@ -10,7 +10,7 @@ ShortestPathHeuristic::ShortestPathHeuristic(size_t source, size_t graph_size, c
     size_t num_of_objectives = adj_matrix.get_num_of_objectives();
     size_t i = 0;
     for (auto node_iter = this->all_nodes.begin(); node_iter != this->all_nodes.end(); node_iter++) {
-        *node_iter = std::make_shared<Node>(i++, std::vector<size_t>(num_of_objectives, 0), std::vector<size_t>(num_of_objectives, MAX_COST));
+        *node_iter = std::make_shared<Node>(i++, std::vector<double>(num_of_objectives, 0), std::vector<double>(num_of_objectives, MAX_COST));
  }
 
     for (int j=0; j < num_of_objectives; j ++){
@@ -19,7 +19,7 @@ ShortestPathHeuristic::ShortestPathHeuristic(size_t source, size_t graph_size, c
 }
 
 
-std::vector<size_t> ShortestPathHeuristic::operator()(size_t node_id) {
+std::vector<double> ShortestPathHeuristic::operator()(size_t node_id) {
     return this->all_nodes[node_id]->h;
 }
